Server
- Use the `bindPort` parameter when configuring the HTTP server
- Updated type of `bindPort` to `ushort`
- Save the newly created `Connection` instance into its own variable
- Do the fiber call later

App

- Updated type of `bindPort` to `ushort`

+++ source/nostril/app.d
@@ -20,7 +20,7 @@ import nostril.server;
 void main()
 {
     string[] bindAddresseds = ["::"];
-    int bindPort = 8082;
+    ushort bindPort = 8082;
 	Server server = new Server(bindAddresseds, bindPort);
     server.startServer();
 }
